Separated the installer code from the main(); Created two examples for Ubuntu 20.04 and Ubuntu 20.04.1 installers
